\name{permutations-package}
\alias{permutations-package}
\alias{permutations}
\docType{package}
\title{
\packageTitle{permutations}
}
\description{
\packageDescription{permutations}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{permutations}
\packageIndices{permutations}

}
\author{
\packageAuthor{permutations}

Maintainer: \packageMaintainer{permutations}
}
\keyword{ package }
\examples{
a <- rperm(10,5)
b <- rperm(10,5)

a*b

inverse(a)


}
