\name{orbit}
\alias{orbit}
\alias{orbit_single}
\title{
Orbits of integers
}
\description{\loadmathjax
Finds the orbit of a given integer
}
\usage{
orbit_single(c1,n1)
orbit(cyc,n)
}
\arguments{
  \item{c1,n1}{In (low-level) function \code{orbit_single()}, a cyclist
    and an integer vector respectively}
  \item{cyc,n}{In (vectorized) function \code{orbit()}, \code{cyc} is
    coerced to a cycle, and \code{n} is an integer vector}
}
\value{

  Given a cyclist \code{c1} and integer \code{n1}, function
  \code{orbit_single()} returns the single cycle containing integer
  \code{n1}.  This is a low-level function, not intended for the
  end-user.

  Function \code{orbit()} is the vectorized equivalent of
  \code{orbit_single()}.  Vectorization is inherited from
  \code{cbind()}.

  The orbit of a fixed point \mjseqn{p} is \mjeqn{\left\lbrace
  p\right\rbrace}{omitted}; the code uses an ugly hack to ensure that
  this is the case.
  
}



\author{Robin K. S. Hankin}
\note{

  Orbits are useful in a more general group theoretic context.  Consider
  a finite group \mjseqn{G} acting on a set \mjseqn{X}, that is

  \mjdeqn{\alpha\colon G\times X\longrightarrow X.}{omitted.}

  Writing \mjeqn{\alpha(g,x)=gx}{omitted}, we define
  \mjeqn{\alpha}{omitted} to be a \dfn{group action} if \mjseqn{ex=x}
  and \mjseqn{g(hx)=(gh)x} where \mjeqn{g,h\in G}{omitted} and
  \mjseqn{e} is the group identity.  For any \mjeqn{x\in X}{omitted} we
  define the \dfn{orbit} \mjseqn{Gx} of \mjseqn{x} to be the set of
  elements of \mjseqn{X} to which \mjseqn{x} can be moved by an element
  of \mjseqn{G}.  Symbolically:

  \mjdeqn{Gx=\left\lbrace gx\colon g\in G\right\rbrace}{omitted}

  Now, we abuse notation slightly.  In the context of permutation
  groups, we consider a fixed permutation \mjeqn{\sigma}{omitted}.  We
  consider the group \mjeqn{G=\left\langle\sigma\right\rangle}{omitted},
  that is, the group \emph{generated by} \mjeqn{\sigma}{omitted}; the
  group action is that of \mjseqn{G} on set \mjeqn{X=\left\lbrace
  1,2,\ldots,n\right\rbrace}{X={1,2,...,n}} with the obvious definition
  \mjeqn{\sigma x=\sigma(x)}{omitted} for \mjeqn{\sigma\in G}{omitted}
  and \mjeqn{x\in X}{omitted}.  This clearly is a group action as
  \mjeqn{\operatorname{id}(x)=x}{id(x)=x} and \mjeqn{\sigma(\mu
  x)=(\sigma\mu)x}{omitted}.

  \mjdeqn{Gx=\bigcup_{i\in\mathbb{Z}}\sigma^i(x)}{omitted}

  Expressing \mjeqn{\sigma}{omitted} in cycle form makes it easy to see
  that the orbit of any element \mjseqn{x} of \mjseqn{X} is just the
  other members in the cycle containing \mjseqn{x}.  For example,
  consider \mjeqn{\sigma=(26)(348)}{omitted} and \mjseqn{x=4}.
  Then

  \mjdeqn{G=\left\langle(26)(348)\right\rangle =
    \bigcup_{i\in\mathbb{Z}}[(26)(348)]^i}{omitted}

  Because we are only interested in the effects on \mjseqn{x=4}, we only
  need to consider the cycle \mjseqn{(348)}: this is the only cycle that
  affects \mjseqn{x=4}, and the \mjseqn{(26)} cycle may be ignored as it
  does not affect element 4.  So

  \mjdeqn{G4=\bigcup_{i\in\mathbb{Z}}(348)^i(4)=\left\lbrace
  3,4,8\right\rbrace}{omitted}

  [observe the slight notational abuse above: \dQuote{\mjseqn{(348)}}
  means \dQuote{the function \mjeqn{f(\cdot)}{f(.)} with
  \mjseqn{f(3)=4}, \mjseqn{f(4)=8}, and \mjseqn{f(8)=3}}].

  }
\seealso{
\code{\link{fixed}}
}
\examples{

orbit(as.cycle("(123)"),1:5)
orbit(as.cycle(c("(12)","(123)(45)","(2345)")),1)
orbit(as.cycle(c("(12)","(123)(45)","(2345)")),1:3)

data(megaminx)
orbit(megaminx,13)

}
\keyword{symbolmath}
